# public/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "TransitOps API Documentation",
    "description": "Comprehensive API specification for the TransitOps fleet operations and dashboard platform. Contains endpoints for vehicles, drivers, trips, maintenance logs, expenses, fuel logs, automated notifications, and dynamic analytics reports.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "/api",
      "description": "TransitOps API Gateway"
    }
  ],
...
  "paths": {
...
    "/vehicles": {
      "get": {
        "tags": ["Vehicles"],
        "summary": "List Vehicles",
        "parameters": [
          { "name": "type", "in": "query", "description": "Filter by type (e.g. Truck, Van, Bus)", "schema": { "type": "string" } },
          { "name": "status", "in": "query", "description": "Filter by status (Available, On Trip, In Shop, Retired)", "schema": { "type": "string" } },
          { "name": "region", "in": "query", "description": "Filter by region (e.g. North, West, South, East)", "schema": { "type": "string" } }
        ],
        "responses": {
          "200": {
            "description": "List of vehicles",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Vehicle" }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Vehicles"],
        "summary": "Create Vehicle",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateVehicleRequest" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Vehicle successfully created",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Vehicle" }
              }
            }
          },
          "422": { "description": "Validation error or registration number already exists" }
        }
      }
    },
...
    "/fuel-logs": {
      "get": {
        "tags": ["Fuel & Expenses"],
        "summary": "List Fuel Logs",
        "responses": {
          "200": {
            "description": "List of fuel logs",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/FuelLog" }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Fuel & Expenses"],
        "summary": "Create Fuel Log",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateFuelLogRequest" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Fuel log entry recorded",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/FuelLog" }
              }
            }
          },
          "422": { "description": "Vehicle not found" }
        }
      }
    },
...
    "/reports/export.pdf": {
      "get": {
        "tags": ["Dashboard & Reports"],
        "summary": "Export Analytics Report as PDF",
        "parameters": [
          { "name": "type", "in": "query", "required": true, "description": "Type of report", "schema": { "type": "string", "enum": ["fuel-efficiency", "utilization", "operational-cost", "roi"] } }
        ],
        "responses": {
          "200": {
            "description": "Downloadable PDF document stream",
            "content": {
              "application/pdf": {
                "schema": { "type": "string", "format": "binary" }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Vehicle": {
        "type": "object",
        "properties": {
          "regNumber": { "type": "string" },
          "name": { "type": "string" },
          "type": { "type": "string" },
          "maxLoadCapacity": { "type": "number" },
          "odometer": { "type": "number" },
          "acquisitionCost": { "type": "number" },
          "status": { "type": "string", "enum": ["Available", "On Trip", "In Shop", "Retired"] },
          "region": { "type": "string" }
        }
      },
      "CreateVehicleRequest": {
        "type": "object",
        "properties": {
          "regNumber": { "type": "string", "example": "MH-12-AB-1234" },
          "name": { "type": "string", "example": "Mahindra Bolero Picker" },
          "type": { "type": "string", "example": "Van" },
          "maxLoadCapacity": { "type": "number", "example": 1500 },
          "odometer": { "type": "number", "example": 25000 },
          "acquisitionCost": { "type": "number", "example": 950000 },
          "status": { "type": "string", "enum": ["Available", "On Trip", "In Shop", "Retired"], "default": "Available" },
          "region": { "type": "string", "example": "West" }
        },
        "required": ["regNumber", "name", "type", "maxLoadCapacity", "odometer", "acquisitionCost", "status", "region"]
      },
...
      "FuelLog": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "liters": { "type": "number" },
          "cost": { "type": "number" },
          "date": { "type": "string", "format": "date" },
          "vehicleReg": { "type": "string" }
        }
      },
      "CreateFuelLogRequest": {
        "type": "object",
        "properties": {
          "vehicleId": { "type": "string", "example": "VAN-05" },
          "liters": { "type": "number", "example": 40 },
          "cost": { "type": "number", "example": 4800 },
          "date": { "type": "string", "format": "date", "example": "2026-07-09" }
        },
        "required": ["vehicleId", "liters", "cost", "date"]
      },
...
    }
  }
}